ADD(UI, editor metadata support): added strongly typed means to provide metadata to UI:
- some refactorings along the way.

Why?
Editor form for creation end update might have differ in terms of fields that are editable, e.g. some filed might not be allowed to change! For this it is not reasonable to have this member in edit object instance, but UI might need to show at least plaintext vale, or disabled field. Metadata can provide both w/o upsert model to have anything that is not editable or needed for given operation.